Initial commit for snf-occi 0.3 with SOI
In version 0.3, a radically different implementation approach is
chosen. The OOI (Openstack OCCI Interface) is used as the primary
tool for processing and operating OCCI calls. The new package, SOI
(Synnefo OCCI Interface) is a wrapper of OOI. It manages the calls
to Synnefo and adjusts OOI to play with Synnefo, through patching.